Moved the canvas commands into their own package unsync_canvas
