+++ unsync_canvas/unsync_canvas/__init__.py
@@ -1,0 +1,17 @@
+__version__ = '0.0.1'
+
+unsync_prefix = 'canvas'
+
+unsync_commands = [
+    'accounts.get_sub_accounts_of_account',
+    'accounts.get_accounts',
+    'accounts.get_courses',
+    'accounts.update_accounts',
+    'enrollment_terms.get_enrollment_terms',
+    'logins.update_user_logins',
+    'logins.get_account_logins',
+    'logins.get_user_logins',
+    'sis_imports.list_imports',
+    'users.get_user_profiles',
+    'users.list_account_users',
+]
